to fazendo o mapa em aqruivo 'igual' em qualquer tamanho de tela
