+++ jogo/classes.py
@@ -1,6 +1,5 @@
 import pygame
 from constantes import *
-from fase_padrao import Fase
 class Jogo:
     def __init__(self):
         pygame.init()
@@ -24,12 +23,43 @@ class Jogo:
     def finaliza(self):
         pygame.quit()
 
-class Fase1(Fase):
+class Fase1:
     def __init__(self):
         self.gera_mapa()
+        self.mapa_img = pygame.image.load('jogo/assets\img\mapa.png')
+    
+    def gera_mapa(self):
+        with open('jogo/mapas/mapa1.txt','w') as mapa1:
+            for y in range(ALTURA_MAPA):
+                if y != 0:
+                    mapa1.write('\n')
+                for x in range(LARGURA_MAPA):
+                    if (x == 0 or y == 0) or (x == LARGURA_MAPA - 1 or y == ALTURA_MAPA - 1):
+                        mapa1.write('1')
+                    # elif LARGURA_MAPA // 2 - 5 <= x <= LARGURA_MAPA // 2 + 5 and ALTURA_MAPA // 2 - 5 <= y <= ALTURA_MAPA // 2 + 5:
+                    #     mapa1.write('1')
+                    else:
+                        mapa1.write('2')
 
     def desenha(self):
-        self.desenha_mapa()
+        JANELA.blit(pygame.transform.scale(self.mapa_img, (LARGURA_MAPA*BLOCO, ALTURA_MAPA*BLOCO)), (MARGEM_X, MARGEM_Y))
+        with open('jogo/mapas/mapa1.txt','r') as mapa1:
+            for y in range(ALTURA_MAPA):
+                linha = mapa1.readline()
+                for x in range(len(linha)):
+                    if linha[x] == '1':
+                        rect = pygame.Rect(x * BLOCO + MARGEM_X, y * BLOCO + MARGEM_Y, BLOCO, BLOCO)
+                        pygame.draw.rect(JANELA, AZUL, rect, 1)
+                    elif linha[x] == '2':
+                        pos_x = x * BLOCO + BLOCO // 2 + MARGEM_X
+                        pos_y = y * BLOCO + BLOCO // 2 + MARGEM_Y
+                        raio = BLOCO // 10
+                        pygame.draw.circle(JANELA, AMARELO_PONTOS, (pos_x, pos_y), raio)
+                    # elif linha[x] == '0':
+                    #     pos_x = x * BLOCO + BLOCO // 2 + MARGEM_X
+                    #     pos_y = y * BLOCO + BLOCO // 2 + MARGEM
+                    #     raio = BLOCO // 2.5
+                    #     pygame.draw.circle(JANELA, AMARELO_PONTOS, (pos_x, pos_y), raio)
 
 if __name__ == '__main__':
     jogo = Jogo()
