+++ jogo/constantes.py
@@ -1,8 +1,15 @@
 import pygame
-#jogo
 TAMANHO_JANELA = (1024,760)
+#mapa
+LARGURA_MAPA = 28
+ALTURA_MAPA = 32
+BLOCO = min(TAMANHO_JANELA[0] // LARGURA_MAPA, TAMANHO_JANELA[1] // ALTURA_MAPA)
+POSICOES_OCUPADAS = []
+
+#jogo
 JANELA = pygame.display.set_mode(TAMANHO_JANELA)
-MARGEM = 80
+MARGEM_X = TAMANHO_JANELA[0]/2 - LARGURA_MAPA*BLOCO/2
+MARGEM_Y = TAMANHO_JANELA[1]/2 - ALTURA_MAPA*BLOCO/2
 
 #cores
 PRETO = (0,0,0)
@@ -10,9 +17,3 @@ BRANCO = (255,255,255)
 AZUL = (0,0,255)
 AMARELO_PONTOS = (255,219,88)
 AMARELO_PAC_MAN = (255,255,0)
-
-#mapa
-LARGURA_MAPA = (TAMANHO_JANELA[0] - MARGEM*2) // 10
-ALTURA_MAPA = (TAMANHO_JANELA[1] - MARGEM*2) // 10
-POSICOES_OCUPADAS = []
-BLOCO = 10
