+++ jogo/mapas/mapa1.txt
@@ -1,60 +1,32 @@
-11111111111111111111111111111111111111111111111111111111111111111111111111111111111111
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-12222222222222222222222222222222222222222222222222222222222222222222222222222222222221
-11111111111111111111111111111111111111111111111111111111111111111111111111111111111111
+1111111111111111111111111111
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1222222222222222222222222221
+1111111111111111111111111111
